moved shared project components
